+++ examples/nextjs/next.config.js
@@ -1,15 +1,8 @@
-// import { webpackPlugin } from '@scrshot/bundler';
-const unplugin = require('@scrshot/bundler').unplugin
-// console.log(webpackPlugin)
+const withScrshot = require('@scrshot/bundler/next')();
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-  webpack: (config) => {
-    // console.log();
-    config.plugins.push(unplugin.webpack({ strip: true }))
-    return config;
-  }
 }
 
-module.exports = nextConfig
+module.exports = withScrshot(nextConfig)