+++ examples/vite-react/vite.config.ts
@@ -1,8 +1,8 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react-swc'
-import { unplugin } from '@scrshot/bundler';
+import Scrshot from '@scrshot/bundler/vite';
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react(), unplugin.vite()],
+  plugins: [react(), Scrshot()],
 })