+++ packages/bundler/src/types.ts
@@ -1,0 +1,11 @@
+export interface Options {
+  /**
+   * Watch images 
+   */
+  watch?: boolean;
+
+  /**
+   * Strip components from production bundle
+   */
+  strip?: boolean;
+};