+++ packages/bundler/tsup.config.ts
@@ -1,0 +1,12 @@
+import type { Options } from 'tsup'
+
+export default <Options>{
+  entryPoints: [
+    'src/*.ts',
+  ],
+  clean: true,
+  format: ['cjs', 'esm'],
+  dts: true,
+  // minify: true,
+  onSuccess: 'npm run build:fix',
+}