+++ packages/config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @scrshot/config
 
+## 0.1.2
+
+### Patch Changes
+
+- Better handling issues and presenting them to user
+
 ## 0.1.1
 
 ### Patch Changes