+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @scrshot/core
 
+## 0.1.2
+
+### Patch Changes
+
+- Remove ScrshotDebug component
+
 ## 0.1.1
 
 ### Patch Changes