+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @scrshot/react
 
+## 0.1.2
+
+### Patch Changes
+
+- Remove ScrshotDebug component
+- Updated dependencies
+  - @scrshot/core@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes